add logger handler
change  path  absolute to  relative

modif test

update yarn

add team and  teamMove

add team-move  to create team

+++ src/users/services/users.service.ts
@@ -1,8 +1,14 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 
-import { encryptPassword } from 'src/common/utils/auth';
+import { Role } from '../../auth/models/roles.model';
+import { PayloadToken } from '../../auth/models/token.model';
+import { encryptPassword } from '../../common/utils/auth';
 import {
   CreateUserDto,
   CreateUserResponseDto,
@@ -11,8 +17,8 @@ import {
   QueryGetUsersDto,
   UpdateUserDto,
   UpdateUserResponseDto,
-} from 'src/users/dto/users.dto';
-import { User } from 'src/users/entities/user.entity';
+} from '../dto/users.dto';
+import { User } from '../entities/user.entity';
 
 @Injectable()
 export class UsersService {
@@ -33,7 +39,22 @@ export class UsersService {
     return users.map((user) => user.toJSON());
   }
 
-  async getUserById(id: string): Promise<GetUserResponseDto> {
+  isUnauthorizedToEditOrSeeUser(jwtUser: PayloadToken, id: string) {
+    return jwtUser.role == Role.User && jwtUser.userId != id;
+  }
+  isUnauthorizedToCreateUserWithRole(jwtUserRole: Role, userRole: Role) {
+    return jwtUserRole != Role.SuperAdmin && userRole != Role.User;
+  }
+
+  async getUserById(
+    jwtUser: PayloadToken,
+    id: string,
+  ): Promise<GetUserResponseDto> {
+    if (this.isUnauthorizedToEditOrSeeUser(jwtUser, id)) {
+      throw new UnauthorizedException(
+        `Unauthorized: You do not have permission to view a user different from '${jwtUser.userId}'.`,
+      );
+    }
     const user = await this.userModel.findById(id).select('-password').exec();
     if (!user) {
       throw new NotFoundException('User not found');
@@ -41,7 +62,15 @@ export class UsersService {
     return user.toJSON();
   }
 
-  async createUser(user: CreateUserDto): Promise<CreateUserResponseDto> {
+  async createUser(
+    jwtUser: PayloadToken,
+    user: CreateUserDto,
+  ): Promise<CreateUserResponseDto> {
+    if (this.isUnauthorizedToCreateUserWithRole(jwtUser.role, user.role)) {
+      throw new UnauthorizedException(
+        `Unauthorized: The user does not have permission to create a user with the role '${user.role}'.`,
+      );
+    }
     const newUser = new this.userModel(user);
     newUser.password = await encryptPassword(newUser.password);
     const saveUser = await newUser.save();
@@ -49,10 +78,20 @@ export class UsersService {
   }
 
   async updateUserById(
+    jwtUser: PayloadToken,
     id: string,
     user: UpdateUserDto,
   ): Promise<UpdateUserResponseDto> {
-    //TODO validate info
+    if (this.isUnauthorizedToEditOrSeeUser(jwtUser, id)) {
+      throw new UnauthorizedException(
+        `Unauthorized: You do not have permission to edit a user different from '${jwtUser.userId}'.`,
+      );
+    }
+    if (this.isUnauthorizedToCreateUserWithRole(jwtUser.role, user.role)) {
+      throw new UnauthorizedException(
+        `Unauthorized: The user does not have permission to update user to role '${user.role}'.`,
+      );
+    }
     if (user.password) {
       user = { ...user, password: await encryptPassword(user.password) };
     }
@@ -66,8 +105,15 @@ export class UsersService {
     return saveUser.toJSON();
   }
 
-  async deleteUserById(id: string): Promise<DeleteUserResponseDto> {
-    //TODO validate if can delete
+  async deleteUserById(
+    jwtUser: PayloadToken,
+    id: string,
+  ): Promise<DeleteUserResponseDto> {
+    if (jwtUser.userId == id) {
+      throw new UnauthorizedException(
+        `Unauthorized: You do not have permission to delete your own user account.`,
+      );
+    }
     const userOld = await this.userModel.findByIdAndRemove(id).exec();
     if (!userOld) {
       throw new NotFoundException('User not found');
